+++ knapsack-optimizer/src/test/java/com/maersk/fse/knapsack/controller/KnapsackOptimizeControllerTest.java
@@ -112,7 +112,7 @@ public class KnapsackOptimizeControllerTest {
 
      @Test
      public void submitProblemTest() throws Exception {
-     Problem request = Problem.builder().capacity(430).weights(new long[] {10, 20, 33}).values(new long[] {10, 3, 30}).build();
+     Problem request = Problem.builder().capacity(430).weights(new int[] {10, 20, 33}).values(new int[] {10, 3, 30}).build();
      String json = objectMapper.writeValueAsString(request);
      Task t1 = new Task();
      when(principal.getName()).thenReturn("user1@test.com");
